% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blirp.R
\docType{package}
\name{blirp-package}
\alias{blirp}
\alias{blirp-package}
\title{blirp: Barrett Lab Internal R Package}
\description{
A package of useful little R functions and tips, written by the members of the Barrett Lab at McGill University.
}
\details{
The Barrett Lab Internal R Package (blirp).
Useful functions and little tricks to help you with R!
}
\author{
\strong{Maintainer}: Timothy THurman \email{timothy.j.thurman@gmail.com}

}
\keyword{internal}
