% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.mem.R
\name{clean.mem}
\alias{clean.mem}
\title{Remove everything from R memory}
\usage{
clean.mem()
}
\arguments{
\item{...}{Runs by itself.}
}
\value{
Deletes everything
}
\description{
Will delete everything in R's memory
}
\note{
Just be careful before deleting your Global environment to be sure this is what you wanted!
}
\examples{
# With vector
z <- c(1,2,3, NA)
rm(list=ls(all=TRUE))
z

}
