% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len_nona.R
\name{len_noNA}
\alias{len_noNA}
\title{Length of a vector, without NAs}
\usage{
len_noNA(vector, ...)
}
\arguments{
\item{vector}{A vector.}

\item{...}{There for compatibility with internal function calls.}
}
\value{
Returns and integer giving the length of the vector, not counting NAs
}
\description{
A function which give the length of an object, but does not count NAs
}
\details{
Not very tested yet. Might cause trouble if called on a matrix or array. When
called on a list, counts each element in the list even if there is an NA
within that element, so be careful.
}
\examples{
# With vector
z <- c(1,2,3, NA)
length(z) # returns 4
len_noNA(z) # returns 3

# with lists
listz <- list(z1 = z, z2 = z, z3 = c(NA, 2, 4), NA)
len_noNA(listz)
length(listz) # returns 4
len_noNA(z) # returns 3

}
