% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_msdn.R
\name{summarise_msdn}
\alias{summarise_msdn}
\title{Calculate the mean, standard deviation, and N simultaneously}
\usage{
summarise_msdn(.df, ...)
}
\arguments{
\item{.df}{An input dataframe, can be used as when piping.}

\item{...}{The}
}
\value{
Returns and integer giving the length of the vector, not counting NAs
}
\description{
This function allows you to express your love of cats.
}
\note{
Not very tested yet. Might cause trouble if called on a matrix. When called on a list, counts each element in the list even if there is an NA within that element
}
\examples{
# With vector
z <- c(1,2,3, NA)
length(z) # returns 4
len_noNA(z) # returns 3

# with lists
listz <- list(z1 = z, z2 = z, z3 = c(NA, 2, 4), NA)
length_noNA(listz)
length(listz) # returns 4
len_noNA(z) # returns 3

}
