% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blirp.R
\docType{package}
\name{blirp-package}
\alias{blirp}
\alias{blirp-package}
\title{blirp: Barrett Lab Internal R Package}
\description{
A package of useful little R functions and tips, written by the members of the Barrett Lab at McGill University.
}
\details{
The Barrett Lab Internal R Package (blirp).
Useful functions and little tricks to help you with R!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/barrettlabecoevogeno/blirp}
  \item Report bugs at \url{https://github.com/barrettlabecoevogeno/blirp/issues}
}

}
\author{
\strong{Maintainer}: Timothy Thurman \email{timothy.j.thurman@gmail.com}

Authors:
\itemize{
  \item Marc-Olivier Beusoleil \email{marc-olivier.beausoleil@mail.mcgill.ca}
}

}
\keyword{internal}
